[molecules] add a script to update molecule/particle files
This is a silly script to modify .common (molecules) and .particle
files.
It can be used to modify (even automatic) another files with similar format too.
For .common files it uses "packages_to_add:" section, and for .particle
"packages:", but any can be used using --section.
Commiting it to this git repository because it could be most useful
here.

When adding an item it aborts if encountered file fragment is not sorted
(can be overridden with --force).
It can sort entries in a "section" too (--sort).
It can delete an item instead of adding (--delete).
Unless --noask is specified, displays diff and asks for confirmation.

usage examples:
	./modify.pl kde.common dev-util/geany
	./modify.pl --sort kde.common dev-util/geany
	./modify.pl --sort kde.common
	./modify.pl --delete kde.common x11-terms/xterm

I don't know if it'll be used. I've made it for fun.
Feel free to (re)move
it to another directory, like scripts/ (I think it fits
better here).