[modify.pl] add support for groups
Now lines without commas with be handled correctly.
Note that I cannot simply split on commas, because the "blocks"
are grouped differently, in a way that is useful for the user during
addition, removal or sorting operations.

There is a limitation on adding an entry: currently it is added to the
first "block" and user is given no choice. This can be fixed easily.